requirements.txt
asn1crypto==0.24.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
cssselect==1.0.3
enum34==1.1.6
feedparser==5.2.1
idna==2.6
ipaddress==1.0.19
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
pyOpenSSL==17.5.0
readability-lxml==0.6.2
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==13.1.2
